moved query into subgraph.ts
